requirements.txt
# requirements.txt

python-dotenv
groq
pandas
numpy
python-dotenv
pyyaml
flask
markdown
openai
tavily-python
PyPDF2